# crest-annotation-openapi/openapi.json
{
 "openapi": "3.0.2",
 "info": {
  "title": "FastAPI",
  "version": "0.1.0"
 },
 "paths": {
...
  "/objects/at/{project_id}": {
   "get": {
    "tags": [
     "object"
    ],
    "summary": "Get Object At",
    "operationId": "get_object_at",
    "parameters": [
     {
      "required": true,
      "schema": {
       "title": "Project Id",
       "type": "string"
      },
      "name": "project_id",
      "in": "path"
     },
     {
      "required": false,
      "schema": {
       "title": "Offset",
       "type": "integer",
       "default": 0
      },
      "name": "offset",
      "in": "query"
     },
     {
      "required": false,
      "schema": {
       "title": "Annotated",
       "type": "boolean"
      },
      "name": "annotated",
      "in": "query"
     },
     {
      "required": false,
      "schema": {
       "title": "Synced",
       "type": "boolean"
      },
      "name": "synced",
      "in": "query"
     },
     {
      "required": false,
      "schema": {
       "title": "Search",
       "type": "string"
      },
      "name": "search",
      "in": "query"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Object"
        }
       }
      }
     },
     "404": {
      "description": "Not found"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/objects/offset/{object_id}": {
   "get": {
    "tags": [
     "object"
    ],
    "summary": "Navigate From",
    "operationId": "navigate_from",
    "parameters": [
     {
      "required": true,
      "schema": {
       "title": "Object Id",
       "type": "string"
      },
      "name": "object_id",
      "in": "path"
     },
     {
      "required": false,
      "schema": {
       "title": "Offset",
       "type": "integer",
       "default": 0
      },
      "name": "offset",
      "in": "query"
     },
     {
      "required": false,
      "schema": {
       "title": "Annotated",
       "type": "boolean"
      },
      "name": "annotated",
      "in": "query"
     },
     {
      "required": false,
      "schema": {
       "title": "Synced",
       "type": "boolean"
      },
      "name": "synced",
      "in": "query"
     },
     {
      "required": false,
      "schema": {
       "title": "Search",
       "type": "string"
      },
      "name": "search",
      "in": "query"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ObjectNavigate"
        }
       }
      }
     },
     "404": {
      "description": "Not found"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/objects/total-of/{project_id}": {
   "get": {
    "tags": [
     "object"
    ],
    "summary": "Get Objects Count",
    "operationId": "get_objects_count",
    "parameters": [
     {
      "required": true,
      "schema": {
       "title": "Project Id",
       "type": "string"
      },
      "name": "project_id",
      "in": "path"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TotalOf"
        }
       }
      }
     },
     "404": {
      "description": "Not found"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/objects/of/{project_id}": {
   "get": {
    "tags": [
     "object"
    ],
    "summary": "Get Objects",
    "operationId": "get_objects",
    "parameters": [
     {
      "required": true,
      "schema": {
       "title": "Project Id",
       "type": "string"
      },
      "name": "project_id",
      "in": "path"
     },
     {
      "required": false,
      "schema": {
       "title": "Annotated",
       "type": "boolean"
      },
      "name": "annotated",
      "in": "query"
     },
     {
      "required": false,
      "schema": {
       "title": "Synced",
       "type": "boolean"
      },
      "name": "synced",
      "in": "query"
     },
     {
      "required": false,
      "schema": {
       "title": "Search",
       "type": "string"
      },
      "name": "search",
      "in": "query"
     },
     {
      "required": true,
      "schema": {
       "title": "Page",
       "minimum": 1.0,
       "type": "integer"
      },
      "name": "page",
      "in": "query"
     },
     {
      "required": true,
      "schema": {
       "title": "Size",
       "type": "integer"
      },
      "name": "size",
      "in": "query"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Paginated_SummaryObject_"
        }
       }
      }
     },
     "404": {
      "description": "Not found"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/objects/all-of/{project_id}": {
   "get": {
    "tags": [
     "object"
    ],
    "summary": "Get All Objects",
    "operationId": "get_all_objects",
    "parameters": [
     {
      "required": true,
      "schema": {
       "title": "Project Id",
       "type": "string"
      },
      "name": "project_id",
      "in": "path"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "title": "Response Get All Objects Objects All Of  Project Id  Get",
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/SummaryObject"
         }
        }
       }
      }
     },
     "404": {
      "description": "Not found"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/objects/id/{object_id}": {
   "get": {
    "tags": [
     "object"
    ],
    "summary": "Get Object",
    "operationId": "get_object",
    "parameters": [
     {
      "required": true,
      "schema": {
       "title": "Object Id",
       "type": "string"
      },
      "name": "object_id",
      "in": "path"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Object"
        }
       }
      }
     },
     "404": {
      "description": "Not found"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/objects/finish/{object_id}": {
   "post": {
    "tags": [
     "object"
    ],
    "summary": "Finish Object",
    "operationId": "finish_object",
    "parameters": [
     {
      "required": true,
      "schema": {
       "title": "Object Id",
       "type": "string"
      },
      "name": "object_id",
      "in": "path"
     },
     {
      "required": false,
      "schema": {
       "title": "Finished",
       "type": "boolean",
       "default": true
      },
      "name": "finished",
      "in": "query"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "404": {
      "description": "Not found"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/objects/lock/{object_id}/{session_id}": {
   "get": {
    "tags": [
     "object"
    ],
    "summary": "Get Lock Status",
    "operationId": "get_lock_status",
    "parameters": [
     {
      "required": true,
      "schema": {
       "title": "Object Id",
       "type": "string"
      },
      "name": "object_id",
      "in": "path"
     },
     {
      "required": true,
      "schema": {
       "title": "Session Id",
       "type": "string"
      },
      "name": "session_id",
      "in": "path"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "404": {
      "description": "Not found"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "object"
    ],
    "summary": "Lock Object",
    "operationId": "lock_object",
    "parameters": [
     {
      "required": true,
      "schema": {
       "title": "Object Id",
       "type": "string"
      },
      "name": "object_id",
      "in": "path"
     },
     {
      "required": true,
      "schema": {
       "title": "Session Id",
       "type": "string"
      },
      "name": "session_id",
      "in": "path"
     },
     {
      "required": false,
      "schema": {
       "title": "Force",
       "type": "boolean",
       "default": false
      },
      "name": "force",
      "in": "query"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "404": {
      "description": "Not found"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/objects/unlock/{object_id}": {
   "post": {
    "tags": [
     "object"
    ],
    "summary": "Unlock Object",
    "operationId": "unlock_object",
    "parameters": [
     {
      "required": true,
      "schema": {
       "title": "Object Id",
       "type": "string"
      },
      "name": "object_id",
      "in": "path"
     },
     {
      "required": false,
      "schema": {
       "title": "Session Id",
       "type": "string"
      },
      "name": "session_id",
      "in": "query"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "404": {
      "description": "Not found"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/objects/uri/{object_id}": {
   "post": {
    "tags": [
     "object"
    ],
    "summary": "Get Image Uri",
    "operationId": "get_image_uri",
    "parameters": [
     {
      "required": true,
      "schema": {
       "title": "Object Id",
       "type": "string"
      },
      "name": "object_id",
      "in": "path"
     },
     {
      "required": false,
      "schema": {
       "title": "Concurreny Limit",
       "type": "integer",
       "default": 3
      },
      "name": "concurreny_limit",
      "in": "query"
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ImageRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "404": {
      "description": "Not found"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/objects/describe/{object_id}": {
   "get": {
    "tags": [
     "object"
    ],
    "summary": "Get Image Description",
    "operationId": "get_image_description",
    "parameters": [
     {
      "required": true,
      "schema": {
       "title": "Object Id",
       "type": "string"
      },
      "name": "object_id",
      "in": "path"
     },
     {
      "required": false,
      "schema": {
       "title": "Concurreny Limit",
       "type": "integer",
       "default": 3
      },
      "name": "concurreny_limit",
      "in": "query"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "404": {
      "description": "Not found"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/objects/cache/{encoded}": {
   "get": {
    "tags": [
     "object"
    ],
    "summary": "Get Cached Image",
    "operationId": "get_cached_image",
    "parameters": [
     {
      "required": true,
      "schema": {
       "title": "Encoded",
       "type": "string"
      },
      "name": "encoded",
      "in": "path"
     },
     {
      "required": false,
      "schema": {
       "title": "Concurreny Limit",
       "type": "integer",
       "default": 3
      },
      "name": "concurreny_limit",
      "in": "query"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "404": {
      "description": "Not found"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/objects/local/{encoded}": {
   "get": {
    "tags": [
     "object"
    ],
    "summary": "Get Local Image",
    "operationId": "get_local_image",
    "parameters": [
     {
      "required": true,
      "schema": {
       "title": "Encoded",
       "type": "string"
      },
      "name": "encoded",
      "in": "path"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "404": {
      "description": "Not found"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/objects/annotations/{object_id}": {
   "get": {
    "tags": [
     "object"
    ],
    "summary": "Get Annotations",
    "operationId": "get_annotations",
    "parameters": [
     {
      "required": true,
      "schema": {
       "title": "Object Id",
       "type": "string"
      },
      "name": "object_id",
      "in": "path"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "404": {
      "description": "Not found"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "object"
    ],
    "summary": "Store Annotations",
    "operationId": "store_annotations",
    "parameters": [
     {
      "required": true,
      "schema": {
       "title": "Object Id",
       "type": "string"
      },
      "name": "object_id",
      "in": "path"
     },
     {
      "required": false,
      "schema": {
       "title": "Session Id",
       "type": "string"
      },
      "name": "session_id",
      "in": "query"
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "title": "Annotation Data",
        "type": "string"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "404": {
      "description": "Not found"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/objects/annotations/pull/{object_id}": {
   "post": {
    "tags": [
     "object"
    ],
    "summary": "Pull Annotations",
    "operationId": "pull_annotations",
    "parameters": [
     {
      "required": true,
      "schema": {
       "title": "Object Id",
       "type": "string"
      },
      "name": "object_id",
      "in": "path"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "404": {
      "description": "Not found"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/objects/annotations/push/{object_id}": {
   "post": {
    "tags": [
     "object"
    ],
    "summary": "Push Annotations",
    "operationId": "push_annotations",
    "parameters": [
     {
      "required": true,
      "schema": {
       "title": "Object Id",
       "type": "string"
      },
      "name": "object_id",
      "in": "path"
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "title": "Annotation Data",
        "type": "string"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "404": {
      "description": "Not found"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/objects/annotations/reset/{object_id}": {
   "post": {
    "tags": [
     "object"
    ],
    "summary": "Reset Annotations",
    "operationId": "reset_annotations",
    "parameters": [
     {
      "required": true,
      "schema": {
       "title": "Object Id",
       "type": "string"
      },
      "name": "object_id",
      "in": "path"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "404": {
      "description": "Not found"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "title": "HTTPValidationError",
    "type": "object",
    "properties": {
     "detail": {
      "title": "Detail",
      "type": "array",
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      }
     }
    }
   },
...
   "ImageRequest": {
    "title": "ImageRequest",
    "type": "object",
    "properties": {
     "thumbnail": {
      "title": "Thumbnail",
      "type": "boolean"
     },
     "width": {
      "title": "Width",
      "type": "integer"
     },
     "height": {
      "title": "Height",
      "type": "integer"
     }
    }
   },
...
   "Object": {
    "title": "Object",
    "required": [
     "id",
     "position",
     "annotation_data"
    ],
    "type": "object",
    "properties": {
     "id": {
      "title": "Id",
      "type": "string"
     },
     "object_uuid": {
      "title": "Object Uuid",
      "type": "string"
     },
     "position": {
      "title": "Position",
      "type": "integer"
     },
     "annotated": {
      "title": "Annotated",
      "type": "boolean"
     },
     "synced": {
      "title": "Synced",
      "type": "boolean"
     },
     "annotation_data": {
      "title": "Annotation Data",
      "type": "string"
     }
    }
   },
   "ObjectNavigate": {
    "title": "ObjectNavigate",
    "required": [
     "id"
    ],
    "type": "object",
    "properties": {
     "id": {
      "title": "Id",
      "type": "string"
     }
    }
   },
...
   "Paginated_SummaryObject_": {
    "title": "Paginated[SummaryObject]",
    "required": [
     "items",
     "pages",
     "page",
     "size"
    ],
    "type": "object",
    "properties": {
     "items": {
      "title": "Items",
      "type": "array",
      "items": {
       "$ref": "#/components/schemas/SummaryObject"
      }
     },
     "pages": {
      "title": "Pages",
      "type": "integer"
     },
     "page": {
      "title": "Page",
      "type": "integer"
     },
     "size": {
      "title": "Size",
      "type": "integer"
     }
    }
   },
...
   "SummaryObject": {
    "title": "SummaryObject",
    "required": [
     "id",
     "position"
    ],
    "type": "object",
    "properties": {
     "id": {
      "title": "Id",
      "type": "string"
     },
     "object_uuid": {
      "title": "Object Uuid",
      "type": "string"
     },
     "position": {
      "title": "Position",
      "type": "integer"
     },
     "annotated": {
      "title": "Annotated",
      "type": "boolean"
     },
     "synced": {
      "title": "Synced",
      "type": "boolean"
     }
    }
   },
   "TotalOf": {
    "title": "TotalOf",
    "required": [
     "total",
     "annotated"
    ],
    "type": "object",
    "properties": {
     "total": {
      "title": "Total",
      "type": "integer"
     },
     "annotated": {
      "title": "Annotated",
      "type": "integer"
     }
    }
   },
   "ValidationError": {
    "title": "ValidationError",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "type": "object",
    "properties": {
     "loc": {
      "title": "Location",
      "type": "array",
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      }
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    }
   },
...
  }
 }
}